+++ tests/pyccel/test_pyccel.py
@@ -720,8 +720,9 @@ def test_multiple_results(language):
     pyccel_test("scripts/runtest_multiple_results.py",
             output_dtype = [int, float, complex, bool, int, complex,
                 int, bool, float, float, float, float, float, float,
-                float, float, float, float, float, float
-                , float, float, float, float, int, int], language=language)
+                float, float, float, float, float, float,
+                float, float, float, float, float, float,
+                float, float, float, float, int, int], language=language)
 
 #------------------------------------------------------------------------------
 def test_elemental(language):